mergereplicates: do not require SampleID column
according to the docs the first column of the metadata
needs to contain the sample ID, but it is not required
that it has a specific name. See for instance the docs
of amp_load.

mergereplicates did require the sample ID column to be
named SampleID which is changed in this PR.